fix: widen bytes field in bridge_test table and add large-IO decode test

+++ internal/metrics/bridge_test.go
@@ -64,7 +64,7 @@ func TestRecordDiskIO(t *testing.T) {
 		devLabel string
 		op       byte
 		opLabel  string
-		bytes    uint32
+		bytes    uint64
 	}{
 		{
 			name:     "disk_write",
